papi.c: Add information to API entry debuging
Thanks to Gary Mohr
-------------------
This patch contains the results of taking a second pass to cleanup
the debug prints in the file papi.c.  It adds entry traces to more
functions that can be called from an application.  It also adds lots
of additional values to the trace entries so that we can see what is
being passed to these functions from the application.