Update HSW presets
Thanks to Gary Mohr
-------------------
Previously we sent updates to the PAPI preset event definitions to improve the
preset cache events on Haswell processors.  In checking the latest source, it
looks like the L1 cache events changes did not get applied quite right.  Here
is a patch to the latest source that will make it the way we had intended.