perf_event[_uncore]: switch to libpfm4 extended masks
Patch due to Gary Mohr, many thanks.
------------------------------------
This patch file contains the changes to make the perf_event and
perf_event_uncore components in PAPI use the libpfm4 extended event
masks.  This adds a number of new masks that can be entered with events
supported by these components. They include a mask 'u' which can be used
control if counting in the user domain should be enabled, a mask 'k'
which does the same for the kernel domain, and a mask 'cpu' which will
cause counting to only occur on a specified cpu.  There are also some
other new masks which may work but have not been tested yet.
------------------------------------