event_info utility: address coverity defect
From Gary Mohr
--------------
This patch corrects a defect reported by Coverity.  The defect reported
that the call to PAPI_enum_cmp_event was setting retval which was never
getting used before it got set again by a call to PAPI_get_event_info.

After looking at the code, I decided that we should not be trying to get
the next event inside a loop that is enumerating masks for the current
event.  It makes more sense to break out of the loop to get masks and
let the outer loop that is walking the events get the next event.
--------------