testlib/test_util.c: Check enum return value
Addresses an issue found by Coverity.
Thanks Gary Mohr,
----------------
The changes in this patch file fixes the only defect in the src/testlib
directory.  The defect reported that the return value from a call to
PAPI_enum_cmp_event was being ignored.  This call to enum events is to
get the first event for the component index passed into this function.
It turns out that the function that contains this code is only ever
called by the overflow_allcounters ctest and it only calls once and
always passes a component index of 0 (perf_event).  So I added code to
check the return value and fail the test if an error was returned.
----------------