ctests/ Address coverity reported defects
Thanks to Gary Mohr for the patch
---------------------------------
he contents of this patch file fix defects reported by Coverity in the
directory 'papi/src/ctests'.

The defect reported in branches.c was that a comparison between
different kinds of data was being done.

The defect reported in calibrate.c was that the variable
'papi_event_str' could end up without a null terminator.

The defects reported in describe.c, get_event_component.c, and
krentel_pthreads.c were that return values from function calls were
being stored in a variable but never being used.

I also did a little clean-up in describe.c.  This test had been failing
for me on Intel NHM and SNBEP but now it runs and reports that it
PASSED.
---------------------------------