Bump NUM_MPX_COUNTERS for linux-perf
Uncore on SNB and newer systems has enough counters to go beyond the 64
array spaces we allocate. This needs a better long term solution.

Reported by Gary Mohr
---------------------
When running on snbep systems with the uncore component enabled, if papi
is configured with debug=yes then the message "Warning!  num_cntrs is
more than num_mpx_cntrs" gets written to stderr.  This happens because
the snbep uncore pmu's have a total of 81 counters and PAPI is set to
only accept a maximum of 64 counters.  This change increases the amount
PAPI will accept to 100 (prevents the warning message from being
printed).
---------------------