Uncore component fix:
By Gary Mohr,
The line that sets exclude_guests in the uncore component is there
because it also there in the core component.  But when I look at it
closer, it is an error in both cases.  I will submit a patch to remove
them and get rid of some commented out code that no longer belongs in
the source.

The uncore events do not support the concept of excluding the host or
guest os so we should never set either bit.  But the core events do
support this concept and libpfm4 provides event masks "mg" and "mh" to
control counting in these domains.  By default if neither is set then
libpfm4 excludes counting in the guest os, if either "mh" or "mg" is
provided as an event mask then it is counted but the other is excluded,
and if both are provided then both are counted.  So when the code forces
the exclude_guest bit to be set it breaks the ability to fully control
what will happen with the masks.

I did not notice the uncore part of this problem when testing on my SNB
system, probably because we use an older kernel which tolerated the bit
being set (or maybe because HSW is handled differently).