perf_event_uncore.c: Check scheduability of events
Patch by Gary Mohr,
-------------------
This patch file adds code to the uncore component to check and make sure
that
the events being opened from an event set can be scheduled by the
kernel.  This
kind of code exists in the core component but was not moved into the
uncore component
because it was felt that it would not be an issue with uncore.  Turns
out the kernel
has the same kind of issues when scheduling uncore events.

The symptoms of this problem will show that the kernel will report that
all events
in the event set were opened successfully but when trying to read the
results one
(or more) of the events will get a read failure.  Seen in the traces and
on stderr
 (if papi configured with debug=yes) as a "short read" error.

The logic is slightly different than what is in the core component
because the events
in the core component are grouped and the ones in the uncore are not
grouped.  When
events are grouped, you only need to enable/disable and read results on
the group leader.
But when they are not grouped you need to do these operations on each
event in the event set.