Adding the NativeAvailValidate patch provided by Gary Mohr.
The problem being addresed is that if there were any problems validating
event masks, then those problems would result in the entire event being
invalid.  The desired action was to test each event mask, and if any
basic event mask can make the event succeed, then the event should be
returned as valid and available.

The solution is to create a large buffer and write events and masks into
this buffer as they are processed, tracking their validity.  At the end
go back and mark the validity of the entire event.  This matches the
standard output of PAPI.