simplify x86 model detection
This patch modifies the way cpu model numbers are detected
on X86 (Intel and AMD).

Instead of having a pfm_xxx_detect() routine per CPU, we factorize
the detection into a common function: pfm_intel_x86_model_detec().
Each PMU needs to provide a pmu->cpu_family and a list of model
numbers to checks. The list is zero-terminated. This eliminates
code duplication and is resulting into a small code size reduction.

Note that for some uncore PMUs, to avoid data duplication, the
code is kept as is.

Signed-off-by: Stephane Eranian <eranian@gmail.com>