Thanks to Gary Mohr for the patch
---------------------

Fix for bugs in PAPI_get_event_info when using the core and uncore components:

PAPI_get_event_info returns incorrect / incomplete information.  The
errors were in how the code handled event masks and their descriptions
so the errors would not lead to program failures, just the possibility
of incorrect labeling of output.

---------------------