Support for the ARM X-Gene processor.
Thanks to William Cohen for this patch.
The events for the Applied Micro X-Gene processor are slightly
different from other ARM processors.  Thus, need to define those
presets for the X-Gene processor.

Note: This patch is not tested by the PAPI team because we do not have
the appropriate hardware.