PPC64 sys_mem_info array size is wrong
Thanks to Carl Love for this patch and the following description:
The variable sys_mem_info is an array of type PAPI_mh_info_t.  It is
statically declared as size 4.  The data for POWER8 is statically
declared in entry 4 of the array which is beyond the allocated array.  The
array should be declared without a size so the compiler will automatically
determine the correct size based on the number of elements being
initialized.  This patch makes the change.