byte_profile.c: PPC64 add support for PPC64 Little Endian to byte_profile.c
Thanks to Carl Love for this patch and the following description:
The POWER 8 platform is Little Endian.  It uses ELF
version 2 which does not use function descriptors.  This
patch adds the needed #ifdef support to correctly compile
the test case for Big Endian or Little Endian.

This patch is untested by the PAPI developers (hardware not easily accessible).