This patch file fixes two problems and adds a performance improvement in "papi_native_avail.
Thanks to Gary Mohr for this patch and the following information:

First it corrects a problem when using the -i or -x options.  The code was putting out
too many event divider lines (lines with all '-' characters).  This has been corrected.

Second it improves the results from "papi_native_avail --validate" when being used
on SNBEP systems.  This system has some events which require multiple masks to be provided
for them to be valid.  The validate code was showing these events as not available because
it did not try to use the event with the correct combination of masks.  The fix checks to see
if a valid form of the event has not yet been found and if so then it tries the event with
specific combinations of masks that have been found to make these events work.

It also adds a check before trying to validate the event with a new mask to see if a valid
form of the event has already been found.  If it has then there is no need to try and validate
the event again.