PPC64, fix L1 data cache read, write and all access equations.
Thanks to Carl Love for this patch and the following documentation:

The current POWER 7 equations for all accesses over counts because it
includes non load accesses to the cache.  The equation was changed to
be the sum of the reads and the writes.  The read accesses to the two
units, can be counted with the single event PM_LD_REF_L1 rather then
counting the events to the two LSU units independently.  The number of
reads to the L1 must be adjusted by subtracting the misses as these
become writes.

Power 8 has four LSU units. The same equations can be used since
PM_LD_REF_L1 counts across all four LSU units.