Fix percent error calculation in ctests/tenth.c
Thanks to Carl Love for this patch and the following documentation:

Do the division first then multiply by 100 when calculating the percent
error.  This will keep the magnitude of the numbers closer.  If you
multiply by 100 before dividing you may exceed the size the representable
number size.

Additionally, by casting the values to floats then dividing we get more
accuracy in the calculation of the percent error.  The integer division
will not give us a percent error less then 1.