This patch file fixes a few problems in the core and uncore components.
The code that managed the native_events array has the ability to grow the array
if it fills up.  The code incorrectly detected when the array needed to be grown
which would lead to modifying memory beyond the end of the array. On some platforms
this resulted in obvious failures but on others it did not show any noticeable problems.
Thanks to Carl Love at IBM for reporting this and identifying the fix included in this
patch file.

When an attempt to encode an event fails, the cpu number in the attribute block was
being left set to zero (block was cleared prior to encode call).  If this event was later
added to an event set, the call to the kernel to open the event would pass a cpu number of zero
making the kernel think that a specific cpu should be used.  This resulted in the kernel returning
an error status which papi mapped to something which did not accurately indicate the problem.
Code has been added to force the cpu number to -1 when an event encode attempt fails.  This
value tells the kernel that a cpu number has not been provided and leads to a much better
papi error code to the user.  The code to do this and the code which sets an invalid config
value when an event encode fails has been moved to where the encode failed flag is set.
This insures that the config value and cpu number in the attribute block will always be left
in the correct state for an event which could not be successfully encoded.

Code was added in a previous commit which fetched the event and event mask descriptions
and saved them in the native event table when an event was allocated.  This allows the
PAPI_get_event_info call to return accurate event names, mask names, event description,
and mask descriptions for all allocated events.  This code contained a 200 byte fixed size
buffer for collecting the event mask descriptions.  If the descriptions did not fit in this
buffer they were just discarded.  This patch increases the buffer size to 1024 bytes and
changes the code to put as much description as will fit into the buffer if the space needed
is greater than 1024 bytes.  On SNBEP systems there were 3 events whose mask descriptions did
not fit in 200 bytes when doing a papi_native_avail list.  They all easily fit into the new
1024 byte buffers.