Patch to reduce cost of using PAPI_name_to_code and add list of supported pmu names to papi_component_avail output
Thanks to Gary Mohr for this patch and its documentation:
---
This patch file contains code to look for either pmu names or component names
on the front of event strings passed to PAPI_name_to_code calls.  If found the
name will be compared to values in each component info structure to see if the
component supports this event.  If the pmu name or component name does not match
the values in the component info structure then there is no need to call this
component for this event.  If the event string does not contain either a pmu name
or a component name then all components will be called.  This reduces the overhead
in PAPI when converting event names to event codes when either component names or
pmu names are provided in the event name string.

To support the above checks, there is also code in this patch to add an array of pmu
names to the component info structure and modifications to the core and uncore components
to save the pmu names supported by each of these components in this new array.

This patch also adds code to the papi_component_avail tool to display the pmu names
supported by each active component.
---