PPC64 fix long event codes
There are eight Power 8 event codes that do not fit into a u32.  The program
used to process the event files and produce the events/power8_events.h file
truncated the event codes to a u32.  The PAPI Power specific struct that
holds the event code is also declared to be a u32. This patch changes the
pme_code field to a u64 and fixes the truncated event codes.

Signed-off-by: Carl Love <cel@us.ibm.com>