Update libpfm4
Current with
------------

commit 7e38ff041cbee37ea6293e165ecf937d8db412b7
Author: Carl Love <cel@us.ibm.com>
Date:   Fri Dec 5 06:43:40 2014 +0100

PPC64 fix long event codes

There are eight Power 8 event codes that do not fit into a u32. The program
used to process the event files and produce the events/power8_events.h file
truncated the event codes to a u32.  The PAPI Power specific struct that
holds the event code is also declared to be a u32. This patch changes the
pme_code field to a u64 and fixes the truncated event codes.

Signed-off-by: Carl Love <cel@us.ibm.com>