User defined events: Enhance PAPI preset events allow user defined events via a user event definition file.
Thanks to Gary Mohr for this patch and its documentation.
--------------------------------------------------------

This patch file enhances the code that processes PAPI preset event definition files (papi_events.csv) so
that it can also now be used to process a user provided event definition file.  PAPI still looks for an
environment variable 'PAPI_USER_EVENTS_FILE' and if found uses its value as the pathname of the user event
definition file to process (same behavior as before). The change is that this is done right after processing
the PAPI preset events rather than at the end of PAPI_library_init (after all components were initialized).
An advantage of using this approach is that now user defined events, like preset events, can define multiple
versions of the same event where each version is customized to a particular hardware platform (or pmu name).

The code which processes preset events was also enhanced in the following ways:

The papi_avail command was updated to also list user defined events in its output.  The papi_avail
help and man page have been updated to include user defined events in the descriptions.  The man page
was also updated to add a "see also" reference to a new 'PAPI_derived_event_files' man page.  A new
'PAPI_derived_event_files' man page to provide the user information about how to build an event
definition file has been added.  This patch file contains both the source file changes (needed by
doxygen) and updated copies of the man pages created by doxygen.

The code now allows both postfix (Reverse Polish Notation) and infix (algebraic) formulas to be entered.
There is a new derived event type 'DERIVED_INFIX' to specify that the formula is provided in the algebraic
format.  The formulas will always be converted to postfix format as part of the event definition processing
so if the user does a 'papi_avail -e <eventName>' later it will always be displayed as a postfix formula.

When defining a new derived event (either preset or user defined), it is now possible to use any already
known native event, preset event or user defined event.  This means that new derived events can be created
as a relationship between other already known (their definitions had to already be processed) derived events.
When derived events are created, there is a list of native events needed by that defined event created and
optionally a formula to compute the derived events value.  If a new derived event is created that depends
on another derived event, then the new event will inherit all the native events used by the event it depends
on and the new derived events formula will be merged with the formula from the event it depends on (if there
was one or if it had an implied formula like derived add or sub).  This means that after event definition
processing completes the event tables inside PAPI always contain the list of all native events needed to
compute the derived events results and a postfix formula that will be used to compute the events result.
So if a user does a 'papi_avail -e <eventName>', the output will show what events PAPI is going to count
and how they will be used to generate the events final value.

A new command 'EVENT' has been added to the code which is intended to be used for user defined events.
It is identical to the existing command 'PRESET' used to define preset events.  They are interchangeable
and both can be used in both preset and user defined event definition files.

The code now allows the user to provide a short and long description for the derived event.  The event
definition commands 'PRESET' and 'EVENT' now support tags of "LDESC" and "SDESC" to identify what is
found in the following string.  This was done the same way as the already supported 'NOTE' tag.

These changes do not support the ability to create #define variables that can then be used in event
definition formulas.  This was supported by the old user event definition code.

These changes delete the existing papi_user_event code (two files that are no longer needed).