Reported by Mark Maurice:
On linux systems without a fortran compiler installed
we get an error when building the PAPI fortran tests.

The reason for the error is that in the Makefile in
the ftests directory the @echo lines start with spaces
instead of tabs. 'make' is fussy about tabs and spaces
and gives a 'missing separator' error if a command
starts with spaces instead of a tab.