Updated CUDA component supporting multiple GPUs and multiple CUDA contexts.
This PAPI CUDA component uses the CUPTI library to get information
about the event counters.

NOTE: To use this PAPI CUDA component, there is a difference from
standard PAPI usage.  When adding PAPI events to the CUDA component,
each event needs to be added from the correct CUDA context.  To
repeat, for each CUDA device, switch to that device and add the events
relevant to that device!

If there is only one CUDA device, then the default context will be
used and things should work as before.