Test case for attaching an eventset to a single CPU rather than a thread (attach_cpu)
Thanks to Gary Mohr for this contribution.

This patch adds a test case to demonstrate how to attach an event set
to a cpu so that the event counts for events in that event set reflect
how many of those events occurred on the attached cpu (instead of the
number of events that occurred in a thread of execution).  See
comments in attach_cpu.c to see how and why to probe with specific
cpus (e.g. ./attach_cpu 3).