Change papi_native_avail to refer to event qualifiers (qual) rather than event masks.
Thanks to Gary Mohr for this patch and the following notes.

This patch file fixes one bug and replaces the term "Unit Mask" and other names
used to identify a unit mask with the term "event qualifier".  This renaming was
done because the term "Unit Mask" has a very specific meaning in the hardware. Many
of the flags and other fields we can now provide with an event to control how it is
counted have nothing to do with the unit masks defined in the manuals provided by the
hardware vendors.

Summary of what changed:

Removed the -d command line argument.  It controlled if units should be displayed in output.  Now we always display units if they are defined (only place I have seen them defined is with rapl events).

Fixed bug when displaying event units.  It was displaying the units information in front of the event name and description.  It now displays the units information after the description.

Renamed the -noumasks argument to -noqual.  This prevents event qualifiers (previously known as unit masks) from being displayed.

Replaced headings "Unit Mask" and "Mask Name" with "Qualifiers" and "Name" (when displaying a single event).