add new perf_events pin event modifier
This patch adds a way to express pinned perf_events event
using libpfm4.

The new modifier is called pinned and it is a boolean modifier.
The default value is false.

Pinning can apply to perf_events per-thread or sytem-wide events.
It requires the event be on the PMU at all times, i.e., event
cannot be multiplexed. The number of active events on the PMU
is CPU specific.

Perf_events supports only one pinned event in group and it
needs to be the leader of the group.

$ perf_examples/task -e inst_retired:any_p:pinned foo

Signed-off-by: Stephane Eranian <eranian@gmail.com>