fix two memory leaks due to miss pfm_release_event()
This patch fixes memory leaks on the the event desc pattrs field
in the following routines:
 - pfm_find_event()
 - pfmlib_raw_pmu_encode()
 - pfmlib_perf_event_encode()

If the cal to pfmlib_parse_event() is successful, then the pattrs
field is allocated and therefore must be freed.

The patch also updates the man page for pfm_get_os_event_encoding()
to remphasize the fact that the fstr field in the arg structure must
be freed by the user when it was requested during the call. Otherwise
memory may be lost.

Reviewed-by: Stephane Eranian <eranian@gmail.com>
Signed-off-by: Gary Mohr <Gary.Mohr@bull.com>