Generating pkg-config files for papi
Thanks to William Cohen for this patch (and to Phil Mucci for the patch review).

Some software makes use of pkg-config
(http://www.freedesktop.org/wiki/Software/pkg-config/) when using
libraries.  pkg-config selects compiler flags and libraries for
compiling user code based on the installation location of the package.
It could make it a bit easier to build other software on papi by
abstracting where the libraries are installed.  Rather than having
some complicated path to the installed library, users could use
"pkg-config --libs --cflags papi" to get that information for the
compile.  If there are multiple versions of papi available on the
machine, the user could get a particular one with something like
"pkg-config --libs --cflags papi-5.4.0".