Clean output from papi_avail tools when there are no user defined events
Thanks to Gary Mohr for this patch.

The changes in this patch improve the output from the papi_avail tool.  It was
printing the user defined events header and a PAPI Error message when no
user defined events existed.  These changes add code in the enum call to return
an error when trying to fetch the first user defined event if no user events are
defined.  This allows the application to detect that no user events are known and
skip printing the user defined event heading.  It also prevents the application
from calling PAPI_get_event_info with a user defined event code that does not exist
which avoids the PAPI Error message.

Also a one line change to modify a debug message type to make the debug output
produced by papi_libpfm4_events.c consistent.