Support PAPI_GRN_SYS granularity for perf component, picking a sane CPU number (patch 1 of 2).
Thanks to William Cohen for this patch and the documentation

The checks in perf_event_open syscall cause a failure when both pid=-1
and cpu=-1.  The perf_event component was passing in pid=-1 and cpu=-1
when PAPI_GRN_SYS was selected.  If possible, the code should pick the
current processor that the command is running so that the permission
check works properly when PAPI_GRN_SYS is used.  The patch also adds a
test fail if PAPI_GRN_SYS unable to add PAPI_TOT_CYC.

Signed-off-by: William Cohen <wcohen@redhat.com>