Example of using LD_PRELOAD with the CUDA component.
A short example of using LD_PRELOAD on a Linux system to intercept
function calls and PAPI-enable an un-instrumented CUDA binary.
Several CUDA events (e.g. SM PM counters) require a CUcontext handle
to be a provided since they are context switched. This means that we
cannot use a PAPI_attach from an external process to measure those
events in a preexisting executable.  These events can only be measured
from within the CUcontext, that is, within the CUDA enabled code we
are trying to measure.  If the user is unable to change the source
code, they may be able to use LD_PRELOAD's ability to trap functions
and measure the events for within the executable.
See src/components/cuda/tests/cuda_ld_preload_example.README for details.