improve handling of hardcoded attributes ifor SNB/IVB/HSW/BDW cycle_activity event and others
This patch changes the behavior of the library for events with hardcoded attributes on Intel
PMUs. Some events have forced attributes, such as cmask or invert or edge. The library prevented
user from overriding these. This is still the case with this patch. However, it accepts passing
the  hardcoded attributes as long as they have the same value. What for? To be consistent and
accept fully-qualified event string as produced by the encoding routines.

The patch also changes all the cycle_activity event tables to allow cmask to be specified on
the cmdline. The point is to also allow cmask to be specified with a raw (hex) umask, such
as in cycle_activity:0x6:c=6. This permits encoding any missing umask or cmask value.

The patch adjusts the validation test suite to work with the new policy.

Signed-off-by: Stephane Eranian <eranian@gmail.com>