fix failing testsuite on Intel Haswell-EP because of RAPL
The validate test suite does not differentiate based on the
host CPU. But for RAPL, some events are not available based
on the host CPU.

On Haswell-EP, the RAPL::ENERGY_CORES event is not supported,
so we cannot systematically test for it. We use ENERGY_PKG
instead.

The RAPL support default to a client RAPL, with CORES, PKG, GPU.
But if the host is Haswell-EP, then RAPL uses, PKG, DRAM only.

The RAPL PMU type is the same for all CPUs, the RAPL event list
varies.

Signed-off-by: Stephane Eranian <eranian@gmail.com>