Reset the number of elements in the perf_pe array
Reading in the tracepoints causes the perf_static_events array to be
cloned and adds elements to the the end of the cloned array.  The
number of elements in the array is tracked by the pme_count field in
perf_event_support.  In addition to reverting to the
perf_static_events array the pme_count field needs to be restored to
the original count of elements in the perf_static_events array. This
is because the library may be initialized multiple times in case
pfm_terminate() is called.

Signed-off-by: William Cohen <wcohen@redhat.com>