Thanks to William Cohen of RedHat for providing a patch that limits shared library soname naming to major version number with following description commit f4ec143eef added versioning to the soname of libpapi.so. However, that previous patch make the soname versioning so detailed that any code linked with that particular version of libpapi.so would only run with that particular version of the library.  If the libpapi.so is replaced with a newer release with minor bug fixes and a minor version number change, the code linked with the older version of the papi library cannot use the newer library because of the change in the shared library soname.
This patch only addresses the issues for some of the possible
platforms and similar patches may be needed for other platforms.

Also, thanks for providing another patch that makes internal infiniband component functions static with foloowing description
Comparing the shared libraries for papi-5.2.0 and papi-5.4.2 with
abidiff showed that some of the inifinband internal functions were not
properly marked as static.  Making the functions static ensures that
the functions can only be used through the proper PAPI interfaces and
avoids the possiblity of some external code directly using them.