Make coretemp internal functions static where possible
As much of the internal of the papi shared library should be hidden.
A number of the internal functions for the perf_event and
coretemp components should be static since they are only used
within the individual component.  Making the functions static allows
the compiler to generate better code and reduce the number of entries
in the PLT (Procedure Link Tables).

Signed-off-by: William Cohen <wcohen@redhat.com>