Fix a bashism found in Makefile.inc
While building on an ubuntu system, hit an error that took
a bit to run down.
	/bin/sh: 1: [: perf_event: unexpected operator
	/bin/sh: 1: [: perf_event_uncore: unexpected operator

This was on an ubuntu system where /bin/sh is actually /bin/dash, and is due to a
bashism in Makefile.inc for the build and clean of cuda_samples.
Swapping out the '==' for an '=' should be safe.

Signed-off-by: Will Schmidt <will_schmidt@vnet.ibm.com>