Only expose the shared libary symbols listed *papi.h files
The shared library should avoid exposing internal symbols of the
library.  This change hides the PAPI's internal symbols when it is
built with libpfm 4. Only the functions in papi.h and the associated
fortran wrapper functions are visible to code using the library.  This
change also makes libpapi.so slightly smaller (29KB for stripped
x86_64 shared libary or about 6%).

Note that a similar patch has been proposed for upstream libpfm4 and
would be needed for the bundled libpfm if papi is being built with the
bundled libpfm4.

Signed-off-by: William Cohen <wcohen@redhat.com>