Limit functions visibility in libpfm
Limiting functions and data structures visibility in libpfm.so
so they are hidden from other other code linked with library can allow
the compiler to generate better quality code and reduce linking
overhead on startup.  Hiding the internal functions and data
structures also allow more flexibility in changing internal
implementation while keeping compatibility with previous versions of
the library.

This patch limits libpfm to making visible the function listed in the
header files it provides.  The llvm clang compiler honor the gcc
visibility option and pragmas.  According to the libabigail tool
abidiff 59 functions and 154 variables were hidden as a result of this
change. The patch reduces the size of the shared library by about 14KB
(0.8%) on x86_64.

Signed-off-by: William Cohen <wcohen@redhat.com>