Force all processors to check event schedulability by reading the counters
There are situations where the perf_event_open syscall will return a
file descriptor for a set of events even when they cannot be scheduled
together.  This occurs on 32-bit and 64-bit ARM processors and MIPS
processors.  This problem also occurs on linux kernels older than
2.6.33 and when the watchdog timer steals a performance counter.  To
check that the performance counters are properly setup PAPI needs to
check that the counter values can be successfully read.  Rather than
trying to avoid this test PAPI will now always do it.

Signed-off-by: William Cohen <wcohen@redhat.com>