Have Fortran test support code report errors more clearly
When a Fortran test called the ftest_skip or ftest_fail the support
code would attempt to print out error strings.  However, this support
code would print out gibberish because the string was not properly
initialized.  There doesn't seem to be a easy way in Fortran to get
the error string, for the time being just print out the error number
and people will need to manually map it back to the string.

Signed-off-by: William Cohen <wcohen@redhat.com>