Update to return PAPI_ENOSUPP when the user cannot count events in kernel space
This is an update to the previous commit changing ENOPERM to ENOSUPP.
On some architectures the ENOPERM was misleading, so ENOSUPP is being
used.

The newer linux kernels now have /proc/sys/kernel/perf_event_paranoid
set to 2 which prevents normal users from counting events in kernel
space.  This will cause a number of the tests to fail when run as a
normal user.  Thanks to William Cohen for pointing out this bug.