tests: multiattach: whitespace/comments/clarifications
digging through the code trying to figure out why it fails with
rdpmc enabled.

it turns out it is seeing wrong running/enabled multiplexing results
even though we aren't multiplexing

tracking this down is a pain because we can't strace/ltrace due to the
code using ptrace to start/stop processes.