fix self_count rdpmc example for X86
This patch fixes the program to return correct counts. The interpretation of hdr->offset
was wrong in the code. The offset is negative and needs to be added to the raw count.

Also change the program to print event deltas which makes more sense as it is simply
measuring the cost of the fib() function.

Signed-off-by: Stephane Eranian <eranian@gmail.com>