perf_event: Avoid unintended libpfm build dependency due to PFM_PMU_MAX enum
This patch is based on one sent by William Cohen <wcohen@redhat.com>

The libpfm pfmlib.h file enumerates the each of performance monitoring
units (PMUs) it can program in pfm_pmu_t type.  The last enum in this
type is PFM_PMU_MAX.  Depending on which specific version of libpfm
being used this specific value could vary.  The problem is that
PFM_PMU_MAX is statically defined in the pfmlib.h file and this was
being used as a loop bounds when iterating to determine which PMUs are
potentially available.  If PAPI was built with an older version of
libpfm and then run with a newer libpfm shared library on a machine
with a larger PFM_PMU_MAX value, none of the PMUs past the smaller
PFM_PMU_MAX used for the the build would be examined or enabled.