perf_event: merge the libpfm4 helper libraries
perf_event and perf_event_uncore had their own almost exactly the
same libpfm4 helper libraries.

Maintaining both was a chore, and it looks like it is possible to just
share one copy.

This does mean that it is now not possible to configure the
perf_event_uncore component without perf_event being enabled, but I am
not sure if that was even possible to begin with.