papi_preset: fix compiler warning
This really confusing warning has been around for a while.

gcc-6.3 reports it in a really odd way:

papi_preset.c: In function ‘check_derived_events’:
papi_preset.c:513:19: warning: ‘__s’ may be used uninitialized in this
function$
    int val = atoi(&subtoken[1]);
                       ^~~~~~~~~~~~
		       papi_preset.c:464:1: note: ‘__s’ was declared here
		        ops_string_merge(char **original, char
*insertion, int replaces, int start_ind$
 ^~~~~~~~~~~~~~~~

But there is no __s variable, or anything to do with where the arrows
are pointing.

gcc-5 gives a better warning:

papi_preset.c: In function ‘check_derived_events’:
papi_preset.c:513:14: warning: ‘tok_save_ptr’ may be used
uninitialized in this$
    int val = atoi(&subtoken[1]);
                  ^
		  papi_preset.c:472:8: note: ‘tok_save_ptr’ was
declared here
  char *tok_save_ptr;

So the thing it seems to be complaining about is that the
*saveptr paramater to strtok_r() is not set to NULL.

According to the manpage I don't think this should be needed?
But I think it should be safe to initialize it anyway.