PAPI power9 event list presets
Here is an initial set of events and changes to help support Power9.

This is based on similar changes that were made for power8 when
initial support was added there.  I've updated the event names to
match what we expect to have in power9, and have done compile/build/
sniff tests.

Signed-off-by: Will Schmidt <will_schmidt@vnet.ibm.com>