ctests: rewrite cycle_ratio test
on Intel platforms PAPI_REF_CYC is a fixed 100MHz cycle count

the test was making the assumption that PAPI_REF_CYC was equal to
the max design freq (not turboboost) and thus as far as I can tell
it never would return the right answer.

This test should probably be moved to validation_tests.