Update to CUDA component to support NVLink.
The CUDA component has been cleaned up and updated to support NVLink.
NVLink metrics can not be measured properly in KERNEL event collection
mode, so the CUPTI EventCollectionMode is transparently set to
CUPTI_EVENT_COLLECTION_MODE_CONTINUOUS when a NVLink metric is being
measured in an eventset.  For all other events and metrics, the CUDA
component uses the KERNEL event collection mode.

A bug in the earlier version was that repeated calls to add CUDA
events were failing because some structures were not cleaned up.  This
should now be fixed.

A new nvlink test was added to the CUDA component tests.